Group the examples by transport and pin the registration rule
The HTTP examples move under examples/Http/, making room for a Tcp/ series
alongside them. Declared namespaces are unchanged, since ExampleRunner matches
ClickHouse.Driver.Examples exactly whatever folder a file sits in.

AGENTS.md carries the checklist for adding an example, because RunAllExamples in
Program.cs is hand-maintained: an example left out of it still compiles, still
lists, and still runs under --filter, so nothing reports that CI never runs it.
DateTimeHandling, AsyncInsert and QBitSimilaritySearch were in that state; they
are registered now, and the three that need a cluster, Cloud credentials or a
JWT are listed as deliberate omissions.

QBitSimilaritySearch read L2DistanceTransposed with GetFloat, which throws
because the server returns Float64.

The README's filter section described matching on the file name; the runner
matches the class name.

+++ examples/Http/DataTypes/Vector_001_QBitSimilaritySearch.cs
@@ -63,7 +63,7 @@ public static class QBitSimilaritySearch
             while (reader.Read())
             {
                 var word = reader.GetString(0);
-                var distance = reader.GetFloat(1);
+                var distance = reader.GetDouble(1);
                 Console.WriteLine($"{word,-12}\t{distance:F6}");
             }
         }
@@ -86,7 +86,7 @@ public static class QBitSimilaritySearch
             while (reader.Read())
             {
                 var word = reader.GetString(0);
-                var distance = reader.GetFloat(1);
+                var distance = reader.GetDouble(1);
                 Console.WriteLine($"{word,-12}\t{distance:F6}");
             }
         }
